added leech log support

+++ bot/helper/mirror_utils/upload_utils/pyrogramEngine.py
@@ -1,22 +1,20 @@
-# Implement By - @anasty17 (https://github.com/SlamDevs/slam-mirrorbot/commit/d888a1e7237f4633c066f7c2bbfba030b83ad616)
-# (c) https://github.com/SlamDevs/slam-mirrorbot
-# All rights reserved
-
 import os
 import logging
 import time
+import threading
 
-from pyrogram.errors import FloodWait
+from pyrogram.errors import FloodWait, RPCError
+from PIL import Image
 
-from bot import app, DOWNLOAD_DIR, AS_DOCUMENT, AS_DOC_USERS, AS_MEDIA_USERS, CUSTOM_FILENAME
-from bot.helper.ext_utils.fs_utils import take_ss, get_media_info
+from bot import app, DOWNLOAD_DIR, AS_DOCUMENT, AS_DOC_USERS, AS_MEDIA_USERS, CUSTOM_FILENAME, LOG_CHANNEL
+from bot.helper.ext_utils.fs_utils import take_ss, get_media_info, get_video_resolution
 
 LOGGER = logging.getLogger(__name__)
 logging.getLogger("pyrogram").setLevel(logging.ERROR)
 
-VIDEO_SUFFIXES = ("MKV", "MP4", "MOV", "WMV", "3GP", "MPG", "WEBM", "AVI", "FLV", "M4V")
+VIDEO_SUFFIXES = ("MKV", "MP4", "MOV", "WMV", "3GP", "MPG", "WEBM", "AVI", "FLV", "M4V", "GIF")
 AUDIO_SUFFIXES = ("MP3", "M4A", "M4B", "FLAC", "WAV", "AIF", "OGG", "AAC", "DTS", "MID", "AMR", "MKA")
-IMAGE_SUFFIXES = ("JPG", "JPX", "PNG", "GIF", "WEBP", "CR2", "TIF", "BMP", "JXR", "PSD", "ICO", "HEIC", "JPEG")
+IMAGE_SUFFIXES = ("JPG", "JPX", "PNG", "WEBP", "CR2", "TIF", "BMP", "JXR", "PSD", "ICO", "HEIC", "JPEG")
 
 
 class TgUploader:
@@ -29,6 +27,7 @@ class TgUploader:
         self.uploaded_bytes = 0
         self.last_uploaded = 0
         self.start_time = time.time()
+        self.__resource_lock = threading.RLock()
         self.is_cancelled = False
         self.chat_id = listener.message.chat.id
         self.message_id = listener.uid
@@ -36,12 +35,12 @@ class TgUploader:
         self.as_doc = AS_DOCUMENT
         self.thumb = f"Thumbnails/{self.user_id}.jpg"
         self.sent_msg = self.__app.get_messages(self.chat_id, self.message_id)
+        self.msgs_dict = {}
+        self.corrupted = 0
+        self.user_settings()
 
     def upload(self):
-        msgs_dict = {}
-        corrupted = 0
         path = f"{DOWNLOAD_DIR}{self.message_id}"
-        self.user_settings()
         for dirpath, subdir, files in sorted(os.walk(path)):
             for filee in sorted(files):
                 if self.is_cancelled:
@@ -51,16 +50,19 @@ class TgUploader:
                 up_path = os.path.join(dirpath, filee)
                 fsize = os.path.getsize(up_path)
                 if fsize == 0:
-                    corrupted += 1
+                    LOGGER.error(f"{up_path} size is zero, telegram don't upload zero size files")
+                    self.corrupted += 1
                     continue
                 self.upload_file(up_path, filee, dirpath)
                 if self.is_cancelled:
                     return
-                msgs_dict[filee] = self.sent_msg.message_id
+                self.msgs_dict[filee] = self.sent_msg.message_id
                 self.last_uploaded = 0
                 time.sleep(1.5)
-        LOGGER.info(f"Leech Done: {self.name}")
-        self.__listener.onUploadComplete(self.name, None, msgs_dict, None, corrupted)
+        if len(self.msgs_dict) <= self.corrupted:
+            return self.__listener.onUploadError('Files Corrupted. Check logs')
+        LOGGER.info(f"Leech Completed: {self.name}")
+        self.__listener.onUploadComplete(self.name, None, self.msgs_dict, None, self.corrupted)
 
     def upload_file(self, up_path, filee, dirpath):
         if CUSTOM_FILENAME is not None:
@@ -81,28 +83,38 @@ class TgUploader:
                     if thumb is None:
                         thumb = take_ss(up_path)
                         if self.is_cancelled:
-                            os.remove(thumb)
+                            if self.thumb is None and thumb is not None and os.path.lexists(thumb):
+                                os.remove(thumb)
                             return
+                    if thumb is not None:
+                        img = Image.open(thumb)
+                        width, height = img.size
+                    else:
+                        width, height = get_video_resolution(up_path)
                     if not filee.upper().endswith(("MKV", "MP4")):
                         filee = os.path.splitext(filee)[0] + '.mp4'
                         new_path = os.path.join(dirpath, filee)
                         os.rename(up_path, new_path)
                         up_path = new_path
                     self.sent_msg = self.sent_msg.reply_video(video=up_path,
-                                                              quote=True,
+                                                              quote=False,
                                                               caption=cap_mono,
                                                               parse_mode="html",
                                                               duration=duration,
-                                                              width=480,
-                                                              height=320,
+                                                              width=width,
+                                                              height=height,
                                                               thumb=thumb,
                                                               supports_streaming=True,
                                                               disable_notification=True,
                                                               progress=self.upload_progress)
+                    try:
+                        app.send_video(LOG_CHANNEL, video=self.sent_msg.video.file_id, caption=cap_mono)
+                    except Exception as err:
+                        LOGGER.error(f"Failed to log to channel:\n{err}")
                 elif filee.upper().endswith(AUDIO_SUFFIXES):
                     duration , artist, title = get_media_info(up_path)
                     self.sent_msg = self.sent_msg.reply_audio(audio=up_path,
-                                                              quote=True,
+                                                              quote=False,
                                                               caption=cap_mono,
                                                               parse_mode="html",
                                                               duration=duration,
@@ -111,35 +123,50 @@ class TgUploader:
                                                               thumb=thumb,
                                                               disable_notification=True,
                                                               progress=self.upload_progress)
+                    try:
+                        app.send_audio(LOG_CHANNEL, audio=self.sent_msg.audio.file_id, caption=cap_mono)
+                    except Exception as err:
+                        LOGGER.error(f"Failed to log to channel:\n{err}")
                 elif filee.upper().endswith(IMAGE_SUFFIXES):
                     self.sent_msg = self.sent_msg.reply_photo(photo=up_path,
-                                                              quote=True,
+                                                              quote=False,
                                                               caption=cap_mono,
                                                               parse_mode="html",
                                                               disable_notification=True,
                                                               progress=self.upload_progress)
+                    try:
+                        app.send_photo(LOG_CHANNEL, photo=self.sent_msg.photo.file_id, caption=cap_mono)
+                    except Exception as err:
+                        LOGGER.error(f"Failed to log to channel:\n{err}")
                 else:
                     notMedia = True
             if self.as_doc or notMedia:
                 if filee.upper().endswith(VIDEO_SUFFIXES) and thumb is None:
                     thumb = take_ss(up_path)
                     if self.is_cancelled:
-                        os.remove(thumb)
+                        if self.thumb is None and thumb is not None and os.path.lexists(thumb):
+                            os.remove(thumb)
                         return
                 self.sent_msg = self.sent_msg.reply_document(document=up_path,
-                                                             quote=True,
+                                                             quote=False,
                                                              thumb=thumb,
                                                              caption=cap_mono,
                                                              parse_mode="html",
                                                              disable_notification=True,
                                                              progress=self.upload_progress)
+                try:
+                    app.send_document(LOG_CHANNEL, document=self.sent_msg.document.file_id, caption=cap_mono)
+                except Exception as err:
+                    LOGGER.error(f"Failed to log to channel:\n{err}")
         except FloodWait as f:
-            LOGGER.info(f)
-            time.sleep(f.x)
-        except Exception as e:
-            LOGGER.error(str(e))
-            self.is_cancelled = True
-            self.__listener.onUploadError(str(e))
+            LOGGER.warning(str(f))
+            time.sleep(f.x * 1.5)
+        except RPCError as e:
+            LOGGER.error(f"RPCError: {e} File: {up_path}")
+            self.corrupted += 1
+        except Exception as err:
+            LOGGER.error(f"{err} File: {up_path}")
+            self.corrupted += 1
         if self.thumb is None and thumb is not None and os.path.lexists(thumb):
             os.remove(thumb)
         if not self.is_cancelled:
@@ -149,9 +176,10 @@ class TgUploader:
         if self.is_cancelled:
             self.__app.stop_transmission()
             return
-        chunk_size = current - self.last_uploaded
-        self.last_uploaded = current
-        self.uploaded_bytes += chunk_size
+        with self.__resource_lock:
+            chunk_size = current - self.last_uploaded
+            self.last_uploaded = current
+            self.uploaded_bytes += chunk_size
 
     def user_settings(self):
         if self.user_id in AS_DOC_USERS:
